+++ libs/content/workspace-ffi/src/apple/api.rs
@@ -1,5 +1,5 @@
 use crate::WgpuWorkspace;
-use egui::{Event, Pos2};
+use egui::{vec2, Event, Pos2};
 use lb_external_interface::lb_rs::Uuid;
 use std::ffi::{c_char, c_void, CStr, CString};
 use std::path::PathBuf;
@@ -47,6 +47,22 @@ pub extern "C" fn set_scale(obj: *mut c_void, scale: f32) {
 pub unsafe extern "C" fn dark_mode(obj: *mut c_void, dark: bool) {
     let obj = &mut *(obj as *mut WgpuWorkspace);
     visuals::init(&obj.context, dark);
+}
+
+/// # Safety
+#[no_mangle]
+pub unsafe extern "C" fn scroll_wheel(obj: *mut c_void, scroll_x: f32, scroll_y: f32) {
+    let obj = &mut *(obj as *mut WgpuWorkspace);
+
+    if obj.raw_input.modifiers.command || obj.raw_input.modifiers.ctrl {
+        let factor = (scroll_y / 50.).exp();
+
+        obj.raw_input.events.push(Event::Zoom(factor))
+    } else {
+        obj.raw_input
+            .events
+            .push(Event::Scroll(vec2(scroll_x, scroll_y)));
+    }
 }
 
 /// # Safety
